requirement.txt
streamlit
PyPDF2
google.generativeai
python-dotenv
langchain
openai
tiktoken